Fix the Windows CI failures the stream-json sweep missed
Two problems, both mine, both invisible on a POSIX host because the code
is behind a windows build tag and never compiled locally. This is the
platform-boundary sweep gap round 2 warned about, realised.

The supervisor's Windows fixture builds its fake agy as a cmd.exe
command line: AgyPath is cmd.exe and Args is {"/c", "<cmdline>"}. The
supervisor now appends "--output-format stream-json" when a job's args
do not already select a format, and with cmd.exe those tokens land at
the END of the command line rather than being parsed as flags, so
"ping -n 60 127.0.0.1 --output-format stream-json" became a usage error
that exited 1 in 30ms. The timeout and cancel the fixture exists to test
never fired. The fixture now terminates its command with "& rem" so the
appended tokens are absorbed by a comment, which is the closest a
cmd.exe stand-in can get to a binary that takes them as flags.

TestRunCapturesOutputAndExit also still asserted the pre-stream-json
contract: it echoed plain text and expected it in out. Plain text is not
NDJSON, so the decoder skipped it and out was empty. It now emits a real
event stream, staged in a file and replayed with `type` rather than
echoed, since quoting JSON braces and quotes through a cmd.exe command
line fails for reasons unrelated to what the test covers.

Separately, TestReadAgyVersionToleratesWaitDelay (added in the round-2
fix wave) writes a #!/bin/sh script with no extension, which Windows
cannot exec. Moved to a posix-tagged file.

Verified with GOOS=windows go vet ./... in addition to the host build,
vet, lint and full test suite.
